Add files via upload
This project leverages advanced machine learning techniques to forecast the likelihood of heart disease in individuals based on a range of health parameters. It is designed to function as an early warning system, aiding healthcare providers and individuals in identifying potential risks and taking proactive measures to prevent heart disease.

Key Features:
1) Data Preprocessing:  Ensuring the quality and readiness of data for model training through meticulous cleaning and preparation.
2) Feature Selection:  Identifying the most significant health indicators contributing to heart disease risk.
3) Model Training:  Employing sophisticated algorithms such as Logistic Regression to develop robust predictive models.
4) Model Evaluation: Assessing the performance of models using comprehensive metrics including accuracy score
5) User Interface: Offering an intuitive interface for users to input health data and obtain predictions.

Technologies Used:
1)  Python
2) Scikit-learn
3) Pandas
4) NumPy
5) Google Colab

This project aspires to facilitate the early detection and prevention of heart disease by providing an accessible and reliable predictive tool to improve patient outcomes and promote proactive healthcare.